==> Collect_data_eicu/01_eicu_diagnosis.ipynb <==
# Filter admissions with specific ICD code

import gc
import gzip
import pandas as pd

# Set the display option to show all columns
pd.set_option('display.max_columns', None)

# CAUTION: Path with the specific ICD codes from a CSV file 
icd = 'CSV\Imports\icd_stroke.csv'

compressed_file_path = r"..\Datasets\eicu-2_0\diagnosis.csv.gz"

df = pd.read_csv(compressed_file_path, compression='gzip')

# Read the specific ICD codes from a CSV file
specific_icd_numbers = pd.read_csv(icd, header=None)

# Split the icd9code column and create multiple rows
df['icd9code'] = df['icd9code'].str.split(', ')
df = df.explode('icd9code')
df = df.dropna(subset=['icd9code'])

# Convert the list of ICD codes to a list of strings
specific_icd_numbers = specific_icd_numbers[0].tolist()

# Filter rows based on desired_icd9codes
filtered_df = df[df['icd9code'].isin(specific_icd_numbers)]

filtered_df = filtered_df.copy()
filtered_df.drop_duplicates(subset='patientunitstayid', keep='first', inplace=True)

# Display the filtered DataFrame
display(filtered_df)

# Export the filtered DataFrame to a CSV file
filtered_df.to_csv('CSV\Exports\o01_eicu_diagnosis.csv', index=False, header=True)

# Free RAM
gc.collect()
