==> Collect_data_eicu/08_eicu_dataframe_concat.ipynb <==
import gc
import gzip
import pandas as pd

# Set the display option to show all columns
pd.set_option('display.max_columns', None)

# lab file 
lab_path = 'CSV\Exports\o04_eicu_lab_final.csv'

# Nurse Charting file 
nursechart_path = 'CSV\Exports\o05_eicu_nurseCharting_final.csv'

# Nurse Charting file 
vitalperiodic_path = 'CSV\Exports\o06_eicu_vitalperiodic.csv'

# Nurse Charting file 
vitalaperiodic_path = 'CSV\Exports\o07_eicu_vitalaperiodic.csv'

lab_df = pd.read_csv(lab_path)

nursechart_df = pd.read_csv(nursechart_path)

vitalperiodic_df = pd.read_csv(vitalperiodic_path)

vitalaperiodic_df = pd.read_csv(vitalaperiodic_path)

"""Change the shape of vital periodic to be the same with other dataframes."""

# Melt the dataframe to reshape it
melted_periodic_df = pd.melt(vitalperiodic_df, id_vars=['uniquepid', 'patientunitstayid', 'observationoffset'],
                    var_name='resultname', value_name='resultvalue')
# Sort the melted dataframe by 'uniquepid'
melted_periodic_df = melted_periodic_df.sort_values(by='uniquepid')

# Remove rows with NaN values in 'resultvalue'
melted_periodic_df = melted_periodic_df.dropna(subset=['resultvalue'])

# Export the filtered DataFrame to a CSV file
melted_periodic_df.to_csv('CSV\Exports\o08_eicu_vitalperiodic_final.csv', index=False, header=True)

"""Change the shape of vital aperiodic to be the same with other dataframes."""

# Melt the dataframe to reshape it
melted_aperiodic_df = pd.melt(vitalaperiodic_df, id_vars=['uniquepid', 'patientunitstayid', 'observationoffset'],
                    var_name='resultname', value_name='resultvalue')
# Sort the melted dataframe by 'uniquepid'
melted_aperiodic_df = melted_aperiodic_df.sort_values(by='uniquepid')

# Remove rows with NaN values in 'resultvalue'
melted_aperiodic_df = melted_aperiodic_df.dropna(subset=['resultvalue'])

# Export the filtered DataFrame to a CSV file
melted_aperiodic_df.to_csv('CSV\Exports\o09_eicu_vitalaperiodic_final.csv', index=False, header=True)

"""lab_df reformation"""

column_to_remove = 'labmeasurenamesystem'

# Use the drop method to remove the specified column
lab_df = lab_df.drop(column_to_remove, axis=1)

lab_df.rename(columns={'labresultoffset': 'offset', 'labname': 'observation',
                       'labresult': 'result'}, inplace=True)

"""nursechart_df reformation"""

nursechart_df.rename(columns={'nursingchartoffset': 'offset',
                              'nursingchartcelltypevalname': 'observation',
                              'nursingchartvalue': 'result'}, inplace=True)

"""melted_periodic_df reformation"""

melted_periodic_df.rename(columns={'observationoffset': 'offset',
                                   'resultname': 'observation',
                                   'resultvalue': 'result'}, inplace=True)

"""melted_aperiodic_df reformation"""

melted_aperiodic_df.rename(columns={'observationoffset': 'offset',
                                    'resultname': 'observation',
                                    'resultvalue': 'result'}, inplace=True)

# concatenate the four dataframes one on top of the other. 
result = pd.concat([lab_df, nursechart_df, melted_periodic_df, melted_aperiodic_df], axis=0, ignore_index=True)

# Export the DataFrame to a CSV file
result.to_csv('CSV\Exports\o10_eicu_four_concat_dataframes.csv', index=False, header=True)

# Free RAM
lab_df = None
result = None
nursechart_df = None
vitalperiodic_df = None
vitalaperiodic_df = None
melted_periodic_df = None
melted_aperiodic_df = None
gc.collect()
